Move to GDExtension (#825)
Move from module to gdextension

+++ kt/godot-library/godot-bootstrap-library/src/main/kotlin/godot/annotation/Notification.kt
@@ -3,4 +3,4 @@ package godot.annotation
 @Register
 @Target(AnnotationTarget.FUNCTION, AnnotationTarget.ANNOTATION_CLASS)
 @Retention(AnnotationRetention.RUNTIME)
-annotation class Notification(val value: Int)
+annotation class Notification(val value: Long)
